refactoring: - refactored stores
