+++ src/index.tsx
@@ -8,6 +8,6 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <App placesCount={777}/>
+    <App/>
   </React.StrictMode>
 );
