+++ src/shared/model/comment/comment.ts
@@ -1,0 +1,9 @@
+import {TUser} from '@/shared/model/user';
+
+export type TComment = {
+  id: string;
+  date: string;
+  user: TUser;
+  comment: string;
+  rating: number;
+}
