+++ src/shared/model/comment/index.ts
@@ -1,0 +1,1 @@
+export * from './comment.ts';
