+++ src/shared/model/user/index.ts
@@ -1,0 +1,1 @@
+export * from './user.ts';
