+++ src/shared/model/user/user.ts
@@ -1,0 +1,5 @@
+export type TUser = {
+  name: string;
+  avatarUrl: string;
+  isPro: boolean;
+}
