+++ src/shared/store/api/index.ts
@@ -1,3 +1,2 @@
 export * from './api.ts';
 export * from './constants.ts';
-export type * from './types.ts';
