--- src/shared/store/api/types.ts
@@ -1,1 +1,0 @@
-
