+++ src/shared/store/index.ts
@@ -1,12 +1,16 @@
 import {configureStore, } from '@reduxjs/toolkit';
 import {offersReducer} from './offers.ts';
+import {favoritesReducer} from './favorites.ts';
 import {useDispatch} from 'react-redux';
 import {authReducer} from './auth.ts';
+import {commentsReducer} from './comments.ts';
 
 export const rootStore = configureStore({
   reducer: {
     offers: offersReducer,
     auth: authReducer,
+    favorites: favoritesReducer,
+    comments: commentsReducer
   }
 });
 
@@ -15,3 +19,6 @@ export const useAppDispatch: () => AppDispatch = () => useDispatch<AppDispatch>(
 export type RootState = ReturnType<typeof rootStore.getState>;
 
 export * from './offers.ts';
+export * from './auth.ts';
+export * from './favorites.ts';
+export * from './comments.ts';
